update :: 대표자만 테이블을 삭제할 수 있도록 하는 검증 추가

+++ src/main/kotlin/team/msg/hiv2/domain/reservation/application/usecase/DeleteReservationUseCase.kt
@@ -1,6 +1,7 @@
 package team.msg.hiv2.domain.reservation.application.usecase
 
 import team.msg.hiv2.domain.reservation.application.spi.ReservationPort
+import team.msg.hiv2.domain.reservation.exception.ForbiddenCommandReservationException
 import team.msg.hiv2.domain.reservation.exception.ReservationNotFoundException
 import team.msg.hiv2.domain.user.application.spi.UserPort
 import team.msg.hiv2.domain.user.domain.constant.UseStatus
@@ -17,6 +18,10 @@ class DeleteReservationUseCase(
         val reservation = reservationPort.queryReservationById(reservationId)
             ?: throw ReservationNotFoundException()
         val users = userPort.queryAllUserByReservation(reservation)
+        val currentUser = userPort.queryCurrentUser()
+
+        if(currentUser.id != reservation.representativeId)
+            throw ForbiddenCommandReservationException()
 
         userPort.saveAll(users.map { it.copy(useStatus = UseStatus.AVAILABLE) })
         reservationPort.deleteReservation(reservation)
